refactor(databricks): group Unity Catalog code under unity_catalog package
Consolidate UC metadata, volume I/O, domain JSON, and MetadataService into
databricks/unity_catalog (mirroring lakebase/) while keeping re-exports on
back.core.databricks for backward compatibility.
