+++ src/back/core/databricks/constants.py
@@ -10,8 +10,6 @@ _LEGACY_PAGE_SIZE = 250
 _LEGACY_MAX_PAGES = 20
 
 API_PREFIX = "/api/2.0"
-FS_FILES_PATH = f"{API_PREFIX}/fs/files"
-FS_DIRS_PATH = f"{API_PREFIX}/fs/directories"
 SQL_WAREHOUSES_PATH = f"{API_PREFIX}/sql/warehouses"
 SCIM_ME_PATH = f"{API_PREFIX}/preview/scim/v2/Me"
 SCIM_USERS_PATH = f"{API_PREFIX}/preview/scim/v2/Users"
