+++ src/back/core/databricks/unity_catalog/__init__.py
@@ -1,0 +1,13 @@
+"""Unity Catalog integration — metadata browsing, volumes, and domain I/O."""
+
+from back.core.databricks.unity_catalog.MetadataService import MetadataService
+from back.core.databricks.unity_catalog.UCDomainIO import UCDomainIO
+from back.core.databricks.unity_catalog.UnityCatalog import UnityCatalog
+from back.core.databricks.unity_catalog.VolumeFileService import VolumeFileService
+
+__all__ = [
+    "MetadataService",
+    "UCDomainIO",
+    "UnityCatalog",
+    "VolumeFileService",
+]
