+++ src/back/objects/domain/SettingsService.py
@@ -1769,7 +1769,7 @@ class SettingsService:
                     "Configure a SQL warehouse under Settings → Databricks first."
                 )
             from back.core.databricks.DatabricksAuth import DatabricksAuth
-            from back.core.databricks.UnityCatalog import UnityCatalog
+            from back.core.databricks.unity_catalog.UnityCatalog import UnityCatalog
 
             auth = DatabricksAuth(host=host, token=token, warehouse_id=warehouse_id)
             uc = UnityCatalog(auth)
@@ -2790,7 +2790,7 @@ class SettingsService:
                     "Configure a SQL warehouse under Settings → Databricks first."
                 )
             from back.core.databricks.DatabricksAuth import DatabricksAuth
-            from back.core.databricks.UnityCatalog import UnityCatalog
+            from back.core.databricks.unity_catalog.UnityCatalog import UnityCatalog
 
             auth = DatabricksAuth(host=host, token=token, warehouse_id=warehouse_id)
             uc = UnityCatalog(auth)
