+++ src/back/objects/registry/scheduler.py
@@ -1279,7 +1279,7 @@ def _run_scheduled_build(
             raise ValidationError("Registry not configured")
 
         from back.core.databricks.DatabricksClient import DatabricksClient
-        from back.core.databricks.VolumeFileService import VolumeFileService
+        from back.core.databricks.unity_catalog.VolumeFileService import VolumeFileService
         from back.core.helpers import resolve_warehouse_id
 
         tm.update_progress(task.id, 5, "Loading domain from registry...")
@@ -1539,7 +1539,7 @@ def _run_scheduled_cohort_materialize(
             raise ValidationError("Registry not configured")
 
         from back.core.databricks.DatabricksClient import DatabricksClient
-        from back.core.databricks.VolumeFileService import VolumeFileService
+        from back.core.databricks.unity_catalog.VolumeFileService import VolumeFileService
         from back.core.helpers import resolve_warehouse_id
         from back.core.graphdb import get_graphdb
         from back.objects.digitaltwin import CohortService
