+++ tests/units/core/test_health.py
@@ -17,7 +17,7 @@ import pytest
 from shared.fastapi import health
 
 # These imports work around __init__.py re-exports that shadow module paths.
-_VFS_MOD = importlib.import_module("back.core.databricks.VolumeFileService")
+_VFS_MOD = importlib.import_module("back.core.databricks.unity_catalog.VolumeFileService")
 # ``health`` imports ``get_lakebase_auth`` from the shared lakebase package,
 # so patch the symbol on that package (not the auth submodule).
 _LBA_MOD = importlib.import_module("back.core.databricks.lakebase")
