+++ tests/units/core/test_metadata_mgt.py
@@ -1,9 +1,9 @@
-"""Tests for back.core.databricks.MetadataService — metadata management."""
+"""Tests for back.core.databricks.unity_catalog.MetadataService — metadata management."""
 
 import importlib
 from unittest.mock import patch
 
-_metadata_service_mod = importlib.import_module("back.core.databricks.MetadataService")
+_metadata_service_mod = importlib.import_module("back.core.databricks.unity_catalog.MetadataService")
 
 from back.core.databricks import (
     MetadataService,
