+++ tests/units/core/test_unity_catalog.py
@@ -4,10 +4,10 @@ import importlib
 import pytest
 from unittest.mock import MagicMock, Mock, patch
 
-_unity_catalog_mod = importlib.import_module("back.core.databricks.UnityCatalog")
+_unity_catalog_mod = importlib.import_module("back.core.databricks.unity_catalog.UnityCatalog")
 
 from back.core.databricks.DatabricksAuth import DatabricksAuth
-from back.core.databricks.UnityCatalog import UnityCatalog
+from back.core.databricks.unity_catalog.UnityCatalog import UnityCatalog
 from back.core.errors import ValidationError
