+++ tests/units/test_user_agent.py
@@ -94,7 +94,7 @@ class TestDatabricksAuthHeaders:
 
 class TestVolumeFileServiceHeaders:
     def test_user_agent_present(self, monkeypatch):
-        from back.core.databricks.VolumeFileService import VolumeFileService
+        from back.core.databricks.unity_catalog.VolumeFileService import VolumeFileService
         from back.core.databricks.DatabricksAuth import DatabricksAuth
 
         auth = MagicMock(spec=DatabricksAuth)
